Stage 2
Implement dynamic unit conversion

The hardcoded unit conversion messages have been replaced with a dynamic conversion system. It now takes user input for a number and a measure, then parses and converts the input. Questions about improperly formatted input or unrecognized units are now dealt with in this system as well.