Update project structure and upgrade Gradle version
This commit refactors the project structure and updates Gradle from version 7.6 to 8.5. It changes the files gradle-wrapper.jar, build.gradle, and settings.gradle, as well as several others. The Kotlin code style has been set as the official one and many files have been relocated to adhere to the new project structure.